Simplify and unify assembly loader provider
